+++ main.py
@@ -144,11 +144,16 @@ def sendEmail(email, result_file) :
     with smtplib.SMTP_SSL(smtp_server, port, context=context) as server:
         server.login(sender_email, PASSWORD)
         server.sendmail(sender_email, receiver_email, text)
-def clearAudios():
+def clearFiles():
     path = os.getcwd()+'/audios/'
-    fileList = os.listdir(path)
-    for file in fileList:
-        os.remove(path+file)
+    path2 = os.getcwd()+'/videos/'
+    if os.path.exists(path) and os.path.exists(path2):
+        fileList = os.listdir(path)
+        for file in fileList:
+            os.remove(path+file)
+        fileList = os.listdir(path2)
+        for file in fileList:
+            os.remove(path2+file)
 if submit_button:
     if name == '' or num_videos == '' or cut_duration == '' or output_file == '' or email == '':
         st.warning('Please enter all the fields')
@@ -159,7 +164,7 @@ if submit_button:
         output_file.split('.')[-1] = 'mp3'
         singer = name.replace(' ', '+')
         num_of_videos = int(num_videos)
-        clearAudios()
+        clearFiles()
         videos = get_videos(singer)
         for video in videos:
             # st.write(video)
